License Update (#245)
* switch to lgpl-v3

* update license for import and packer

* update engine license

* update read me

+++ update_license.py
@@ -1,14 +1,19 @@
 import os
 import subprocess
+import logging
+
+# Configure logging
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
 
 # Define the directory to start the search from
-root_directory = './ToolKit/Editor'
+root_directory = './ToolKit'
 
 # Define the file extensions to consider
 file_extensions = ['.h', '.cpp']
 
-# Define the license text to insert
-license_text = '''/*
+# Define the existing and new license text
+existing_license_text = '''/*
  * MIT License
  *
  * Copyright (c) 2019 - Present Cihan Bal - Oyun Teknolojileri ve Yazılım
@@ -33,6 +38,14 @@ license_text = '''/*
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
+
+'''
+new_license_text = '''/*
+ * Copyright (c) 2019-2024 OtSofware
+ * This code is licensed under the GNU Lesser General Public License v3.0 (LGPL-3.0).
+ * For more information, including options for a more permissive commercial license,
+ * please visit [otyazilim.com] or contact us at [info@otyazilim.com].
+ */
  
 '''
 
@@ -41,15 +54,25 @@ for root, dirs, files in os.walk(root_directory):
     for file in files:
         if file.endswith(tuple(file_extensions)):
             file_path = os.path.join(root, file)
+            logger.info(f"Processing file: {file_path}")
+
             with open(file_path, 'r') as f:
                 content = f.read()
 
-            # Insert the license text at the beginning of the file
-            modified_content = license_text + content
+            # Check if the file contains the existing license text
+            if existing_license_text in content:
+                logger.info(f"Existing license text found in {file_path}")
+                # Replace existing license text with the new one
+                modified_content = content.replace(existing_license_text, new_license_text)
 
-            # Write the modified content back to the file
-            with open(file_path, 'w') as f:
-                f.write(modified_content)
-            
-            # Run clang-format to format the file
-            subprocess.run(['clang-format', '-i', file_path])
+                # Write the modified content back to the file
+                with open(file_path, 'w') as f:
+                    f.write(modified_content)
+
+                logger.info(f"License text replaced in {file_path}")
+
+                # Run clang-format to format the file
+                subprocess.run(['clang-format', '-i', file_path])
+                logger.info(f"File formatted: {file_path}")
+
+logger.info("Script execution completed.")
